% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpha_plot.R
\name{alpha_plot}
\alias{alpha_plot}
\title{Build boxplots of alpha diversity}
\usage{
alpha_plot(
  taxa_table = NULL,
  metadata = NULL,
  test_metadata = NULL,
  taxa_level = "species",
  method = "wilcoxon",
  xlab_direction = 1,
  palette_group = c("red", "blue", "orange", "green")
)
}
\description{
Build boxplots of alpha diversity
}
\examples{


}
\keyword{alpha-diversity,}
\keyword{boxplots}
