% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_pathway.R
\name{format_pathway}
\alias{format_pathway}
\title{Parse metagenomics outputs from kraken, metaplan, humann}
\usage{
format_pathway(taxa_file = NULL, sep = "\\t")
}
\description{
Parse metagenomics outputs from kraken, metaplan, humann
}
\examples{

}
\keyword{format,}
\keyword{wgs}
