% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_wgs.R
\name{format_wgs}
\alias{format_wgs}
\title{Parse metagenomics outputs from kraken, metaplan, humann}
\usage{
format_wgs(taxa_file = NULL, sep = "\\t", method = "kraken2")
}
\description{
Parse metagenomics outputs from kraken, metaplan, humann
}
\examples{

}
\keyword{format,}
\keyword{wgs}
