% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_corplot.R
\name{meta_corplot}
\alias{meta_corplot}
\title{Scatter plots of individual taxa with selected metadata
colored by selected groups
with correlation stats}
\usage{
meta_corplot(
  taxa_table = NULL,
  metadata = NULL,
  test_metadata = NULL,
  col_metadata = NULL,
  page = 1,
  log_norm = T,
  fdr_cutoff = 0.1,
  cor_method = "spearman",
  taxa_shown = "",
  palette_group = c("red", "blue", "orange", "green")
)
}
\description{
Scatter plots of individual taxa with selected metadata
colored by selected groups
with correlation stats
}
\examples{

}
\keyword{correlation}
\keyword{taxa,}
