% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_format.R
\name{meta_format}
\alias{meta_format}
\title{Parse metadata file}
\usage{
meta_format(metadata = NULL, metadata_sep = ",", meta_sample_name_col = 0)
}
\description{
Parse metadata file
}
\examples{


}
\keyword{format,}
\keyword{meta}
