% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_test.R
\name{stat_test}
\alias{stat_test}
\title{Statistical tests
wilcoxon, t-test, anova, kruskal-wallis}
\usage{
stat_test(
  taxa_table = NULL,
  metadata = NULL,
  test_metadata = NULL,
  method = "wilcoxon"
)
}
\description{
Statistical tests
wilcoxon, t-test, anova, kruskal-wallis
}
\examples{

}
\keyword{statistical}
\keyword{tests}
