% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_subset.R
\name{table_subset}
\alias{table_subset}
\title{Subset data}
\usage{
table_subset(
  taxa_table = NULL,
  metadata = NULL,
  stratify_by_metadata = "",
  stratify_by_value = "",
  prevalence_cutoff = 0.1,
  abundance_cutoff = 0,
  taxa_file = F
)
}
\description{
Subset data
}
\examples{

}
\keyword{subset}
