% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxa_boxplot.R
\name{taxa_boxplot}
\alias{taxa_boxplot}
\title{Boxplots of individual taxa with stats}
\usage{
taxa_boxplot(
  taxa_table = NULL,
  metadata = NULL,
  test_metadata = NULL,
  one_level = F,
  fdrs = NULL,
  log_norm = T,
  cutoff = 0.1,
  xlab_direction = 1,
  page = 1,
  palette_group = c("red", "blue", "orange", "green"),
  taxa_shown = ""
)
}
\description{
Boxplots of individual taxa with stats
}
\examples{

}
\keyword{boxplot,}
\keyword{individual}
\keyword{stats}
\keyword{taxa,}
