% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_asv.R
\name{taxa_edit}
\alias{taxa_edit}
\title{Parse various 16S outputs
biom file, multiple biom files at different taxonomic levels, one or separate tsv or csv files}
\usage{
taxa_edit(list1)
}
\description{
Parse various 16S outputs
biom file, multiple biom files at different taxonomic levels, one or separate tsv or csv files
}
\examples{

}
\keyword{16s}
\keyword{format,}
