% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mds_plot.R
\name{mds_plot}
\alias{mds_plot}
\title{PCoA plots and PERMANOVA tests}
\usage{
mds_plot(
  taxa_table = NULL,
  metadata = NULL,
  test_metadata = NULL,
  taxa_level = "genus",
  method_mds = "pcoa",
  xaxis = 1,
  yaxis = 2,
  distance_type = "bray",
  palette_group = c("red", "blue", "orange", "green")
)
}
\description{
PCoA plots and PERMANOVA tests
}
\examples{


}
\keyword{PCoA,}
\keyword{PERMANOVA}
